chromium: Use the right bb.utils call in do_check_fdlimit_for_linking
`bb.utils.contains()` requires both "ld-is-gold" and "ld-is-lld" to be
present in `DISTRO_FEATURES`, which in practice means the rlimit check will
happen even when one is not using ld.bfd.

Switch to `bb.utils.contains_any()`, which tests for the presence of any of
the arguments, as we originally intended.

Signed-off-by: Raphael Kubo da Costa <raphael.kubo.da.costa@intel.com>