firefox: Import latest OpenMAX IL 1.1.2 headers
OpenMAX IL 1.1.2 header files were released twice on the same version,
and Firefox 68 includes the OLD headers.

The main change between the two versions is definition of fixed-sized
integers.  For example, the old version defines OMX_U32 as unsigned long,
but the new version defines it as uint32_t.

The definition of the old version is apparently wrong, because there is
no guarantee that unsigned long is 32-bit length.  Actually unsigned
long is 64-bit length on ARM64 and this mismatch cause a problem when
using OpenMAX on RZ/G2 series SoCs.

Signed-off-by: INAJIMA Daisuke <inajima@soum.co.jp>